Solve December23 part 2
Unfortunately, I was unable to actually solve this by actually using
the given data as input.

I'm not sure you were supposed to, but it still bugs me.

Comparing others' solutions in the AoC reddit sub, I noticed that
apparently all input files are the same except for the very first
line assigning different initial values to b. In particular, the
offset to c (the target value) is always 17000 and the step is always
17. That means I could have parsed the input file for said initial
value of b (or going further, also read the difference and step
values). But I don't really see the point in that, given that I'd
still have to assume the rest of the input data progressed in the
same way.

Instead, I opted to implement the functionality (counting non-prime
numbers between two numbers) with 3 parameters for the start, target
and step values, which at least partly satisfies my desire for a
generic solution.